[edit] manage command

+++ cmd/manage.go
@@ -48,33 +48,60 @@ var manageCmd = &cobra.Command{
 			fmt.Println("No matching meeting with the given theme.")
 			os.Exit(1)
 		}
-				var participator []string
-		fmt.Println("Participators:")
-		for i, v := range meetingList[0].GetParticipators() {
-			participator = append(participator, v)
-			fmt.Printf("%d. %s\n", i, v)
-		}
-		fmt.Println("Please input the number you want to remove: ")
-		var inputNums string
-		var tmp int
-		fmt.Scanf("%s", &inputNums)
-		fmt.Scanf("%d", &tmp)
-		chosenList := strings.Split(" ", inputNums)
-		var toBeRemovedParticipators []string
-		for _, v := range chosenList {
-			num, err := strconv.Atoi(v)
-			if err != nil {
-				fmt.Fprintln(os.Stderr, "error: Invalid input")
-				os.Exit(1)
+
+		// delete users
+		if isDelete {
+			var participator []string
+			fmt.Println("Participators:")
+			for i, v := range meetingList[0].GetParticipators() {
+				participator = append(participator, v)
+				fmt.Printf("%d. %s\n", i, v)
 			}
-			toBeRemovedParticipators = append(toBeRemovedParticipators, participator[num])
-		}
-		for _, v := range toBeRemovedParticipators {
-			ok := Service.DeleteParticipatorByTitle(name, meetingName, v)
-			if ok {
-				fmt.Printf("%s was removed.\n", v)
-			} else {
-				fmt.Printf("%s can not be removed.\n", v)
+			fmt.Println("Please input the number you want to remove: ")
+			var inputNums string
+			var tmp int
+			fmt.Scanln(&inputNums)
+			fmt.Scanf("%d", &tmp)
+			chosenList := strings.Split(inputNums, " ")
+			var toBeRemovedParticipators []string
+			for _, v := range chosenList {
+				num, err := strconv.Atoi(v)
+				if err != nil {
+					fmt.Fprintln(os.Stderr, "error: Invalid input")
+					os.Exit(1)
+				}
+				toBeRemovedParticipators = append(toBeRemovedParticipators, participator[num])
+			}
+			for _, v := range toBeRemovedParticipators {
+				ok := Service.DeleteParticipatorByTitle(name, meetingName, v)
+				if ok {
+					fmt.Printf("%s was removed.\n", v)
+				} else {
+					fmt.Printf("%s can not be removed.\n", v)
+				}
+			}
+		} else {
+			// add users
+			fmt.Println("You can choose some of them to add to your meeting:")
+			userList := Service.ListAllUsers()
+			for i, v := range userList {
+				fmt.Printf("%d. %s\n", i + 1, v.GetUserName())
+			}
+			fmt.Println("Please input the number of users you want to add(separate with blank): ")
+			var userNums string
+			fmt.Scanln(&userNums)
+			userNumList := strings.Split(userNums, " ")
+			for _, v := range userNumList {
+				i, ok := strconv.Atoi(v)
+				if ok != nil || i >= len(userList) {
+					fmt.Fprintln(os.Stderr, "error: Invalid input.")
+					os.Exit(0)
+				}
+				if Service.AddParticipatorByTitle(name, meetingName, userList[i - 1].GetUserName()) {
+					fmt.Printf("%s was added.\n", userList[i - 1].GetUserName())
+				} else {
+					fmt.Printf("%s can not be added.\n", userList[i - 1].GetUserName())
+				}
 			}
 		}
 	},
@@ -93,4 +120,5 @@ func init() {
 	// is called directly, e.g.:
 	// manageCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 	manageCmd.Flags().StringVarP(&meetingName, "name", "", "", "the name of meeting to be managed")
+	manageCmd.Flags().BoolVarP(&isDelete, "", "d", false, "Delete a meeting")
 }
